Removes the generateMetadata from layout and moved to home page instead
